Create SLinked list
Implement the following in the template version of the class SLinked list:
a. Add to the back and remove from back of the list
i. void addBack(….);
ii. E removeBack(…);
b. Insert elements in a sorted list i.e., find the position where an element fits
and add it to the list. Note you must call the addBack and addFront
functions if you insert in the front or rare of the list (same for remove
functions)
i. void insert(E value)
ii. void delete(E value)